backend/app/services/bambulab.py
"""Bambu Lab 3D printer client stub.

This module provides a placeholder for MQTT-based communication with
Bambu Lab printers. The actual implementation requires the printer's
IP address, access code, and serial number.
"""

import json
import logging
from typing import Any

logger = logging.getLogger(__name__)


class BambuLabClient:
    """Client for communicating with Bambu Lab printers via MQTT."""

    def __init__(
        self,
        ip_address: str,
        access_code: str,
        serial_number: str,
        port: int = 8883,
    ):
        self.ip_address = ip_address
        self.access_code = access_code
        self.serial_number = serial_number
        self.port = port
        self._connected = False
        self._client = None

    def connect(self) -> None:
        """Establish MQTT connection to the printer.

        TODO: Implement actual MQTT connection using paho-mqtt.
        The connection uses TLS with the printer's access code as password.
        Topic: device/{serial_number}/report
        """
        try:
            import paho.mqtt.client as mqtt

            self._client = mqtt.Client(
                callback_api_version=mqtt.CallbackAPIVersion.VERSION2,
                client_id=f"homelab_manager_{self.serial_number}",
                protocol=mqtt.MQTTv311,
            )
            self._client.username_pw_set("bblp", self.access_code)
            self._client.tls_set()
            self._client.tls_insecure_set(True)

            logger.info(f"Bambu Lab client configured for {self.ip_address}:{self.port}")
            # Uncomment to actually connect:
            # self._client.connect(self.ip_address, self.port)
            # self._client.loop_start()
            # self._connected = True
        except ImportError:
            logger.warning("paho-mqtt not installed, Bambu Lab integration unavailable")
        except Exception as e:
            logger.error(f"Failed to connect to Bambu Lab printer: {e}")

    def disconnect(self) -> None:
        """Disconnect from the printer."""
        if self._client and self._connected:
            self._client.loop_stop()
            self._client.disconnect()
            self._connected = False

    def get_status(self) -> dict[str, Any]:
        """Get current printer status.

        Returns:
            Dict with printer status information.
        """
        # Placeholder response
        return {
            "connected": self._connected,
            "ip_address": self.ip_address,
            "serial_number": self.serial_number,
            "status": "unknown",
            "message": "Bambu Lab integration is a placeholder. Connect to get live data.",
        }

    def send_command(self, command: dict[str, Any]) -> bool:
        """Send a command to the printer via MQTT.

        Args:
            command: Command dict to publish.

        Returns:
            True if command was sent successfully.
        """
        if not self._connected or not self._client:
            logger.warning("Not connected to printer, cannot send command")
            return False

        topic = f"device/{self.serial_number}/request"
        payload = json.dumps(command)
        result = self._client.publish(topic, payload)
        return result.rc == 0
