backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
alembic>=1.13.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
anthropic>=0.18.0
openpyxl>=3.1.0
httpx>=0.27.0
paho-mqtt>=2.0.0
python-dateutil>=2.8.0
aiofiles>=23.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
